moved getFiles function to RootTools
- removes a dependency to H2TauTau in RootTools
- in H2TauTau, cleaned up imports of this function, and obsolete files.
