+++ CMGTools/H2TauTau/python/proto/samples/data_diTau.py
@@ -1,7 +1,6 @@
 import copy
 import os
 import CMGTools.RootTools.fwlite.Config as cfg
-from   CMGTools.H2TauTau.proto.samples.getFiles import getFiles
 
 H2TauTauPackage = '/'.join( [ os.environ['CMSSW_BASE'],
                               'src/CMGTools/H2TauTau' ] )
