+++ CMGTools/H2TauTau/python/proto/samples/run2012/mc_basicEWK.py
@@ -1,7 +1,6 @@
 import copy
 import os
 import CMGTools.RootTools.fwlite.Config as cfg
-from   CMGTools.H2TauTau.proto.samples.getFiles import getFiles
 
 
 DYJets = cfg.MCComponent(
